update v1 example
